pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gmail-backup"
version = "0.1.0"
description = "Gmail attachment backup tool to S3"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.34.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
gmail-backup = "gmail_backup.backup:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]